requirements.txt
fastapi
uvicorn[standard]
PyJWT[crypto]